POC for soundscape-audio C++ library using FMOD
The long term aim is to have a C++ library which provides an API for
Kotlin to create and position audio beacons and other sounds. This proof
of concept adds the library and initializes it from the Kotlin code.
The current behaviour is that it spins up a thread which continually
plays audio with the listener moving around in 3D space. The audio
switches between two different beacon sounds. The aim is to prove that
the FMOD audio engine can do what we want prior to implementing the full
API.