Move audio main loop out of C++ and into Kotlin
Instead of running a main loop on a timer in the C++ audio engine, move
it into Kotlin where it can be driven by the orientation callback.
That's the main event that affects the 3D audio and so it's a sensible
option to update the audio positioning.
Also tidied up startup and shutdown.