Kotlin classes with native backends created and audio tidy up
The main change here is that we can now create AudioBeacon Kotlin
objects in the LocationService and that talks to the C++ library and
creates native beacons. This seems to be the more correct way of doing
things and has allowed tidying up of AudioEngine and AudioBeacon
functionality. The result is that not only do the beacons still work,
but the service can be stopped and restarted reliably and I believe
everything is correctly locked.