More refactoring of audio for initial TextToSpeech
The TextToSpeech engine is initialized as part of the Kotlin
AudioEngine. Each time a TextToSpeech beacon is created in Kotlin it
rustles up a pair of sockets and synthesizes the speech in one end. The
other end is passed into the C++ engine. It reads the data out and
positions it in the same way that the regular Beacon works.
The main complexity is trying to spot the end of the data within the C++
engine. Currently it's a bit of a hack, though it does seem to work.
When the data is finished the beacon is destroyed in C++. Because the
TextToSpeech audio has a GPS location, it's rendered in 3D space at that
location and so turning the device during playback moves it...