Improve text to speech socket handling
Switch socket in C++ to non-blocking and add a timeout if no data is
received for over 10 PCM data callbacks. On the Kotlin side add an
UtteranceProgressListener to follow the progress of the text to speech
generation. For reasons unknown onDone is never called, only onStart.
This means that the Kotlin sockets are destroyed rather lazily, but at
least they are destroyed.