Add support for different styles of audio beacon
BeaconDescriptor now includes lists of assets and at which angles they
should be played. All of the iOS beacon assets have been added and a few
have been added to GetBeaconDescriptor for testing purposes. We can add
the rest in when we add the ability to switch the Beacon used dynamically.