Change min  SDK from 28 to 30
The minimum SDK of 28 didn't include the TextToSpeech API being used,
so this change ups it to 30. Unfortunately that rules out any of my old
phones and this moves from a 2018 release to 2020. I think this is as new
as we should go, and later we can investigate reducing it back down.